+++ CHANGELOG.md
@@ -8,4 +8,15 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+* A package `github.com/fhofherr/acmeproxy/pkg/acme` which enables
+  `acmeproxy` to perform an HTTP01 challenge.
+  ([#2](https://github.com/fhofherr/acmeproxy/issues/2))
+* A `github.com/fhofherr/acmeproxy/pkg/acmetest` package containing
+  a type `Pebble`. It represents an instance of the
+  [pebble](https://github.com/letsencrypt/pebble) test server.
+* A local test environment using `docker-compose`. It runs `pebble` and
+  allows to execute any integration tests locally.
+
 [Unreleased]: https://github.com/fhofherr/leproxy